library-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
oracledb==2.1.1
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1